Update Player.cs
In FireLaser( ) method, changed Input.GetKeyDown(KeyCode.Space) to KeyCode.Return. Was having issues with having the player move diagonally AND firing simultaneously. It wouldn't do both. Opened trouble ticket with Tom and discovered that the keyboard many times shares input signal between the space bar and the movement arrows. So changed from KeyCode.Space to KeyCode.Return to use the Enter key for firing instead and it worked.

Adjusted the position wrapping in the y axis and the constraints on the x axis to make the game a side scroller with a horizontal orientation as opposed to a vertical orientation.