Laser Prefabs
Created 2 additional laser prefabs, laserPrefabLeft and laserPrefabRight.
These are instantiated at a left offset and a right offset to the Player and also the player instantiates 3 laser instantiations for 3 center lasers, 3 lasers on the left and 3 lasers on the right.
Later, will create a rotation so that the left and right lasers fire at an angle from the player.