Laser Prefabs
Updated laser prefabs. Instead of instantiating left, right and center lasers for each group separately, created an empty object that holds the 3 lasers. Inside of Player.cs, when firing, it instantiates the Container Object that holds the 3 lasers attached as children.

Instantiate 3 at the same time, with an offset along the x axis for each object.

The Lazer.cs script was attached to each individual laser object. So, I detached the script from the individual laser objects and attached it to the parent LaserContainer. That's because while they were attached to the individual child objects, the destroy object wasn't executing as desired. Destroy object ensures object is destroyed once it's no longer needed, once it's gone outside the viewable screen area. Now, the parent container controls the speed and the Destroy method.