Spawn Manager Update
Updated SpawnManager.cs and Player.cs so that once the player dies, the SpawnManager stops spawning enemies.

Did this by creating a spawnManager variable inside of Player.cs that grabs the spawnManager component of SpawnManager.

Created a method inside of SpawnManager.cs called StopSpawning() that sets a ContinueSpawning bool to false.

The while(true) loop that instantiates the enemy, I changed the while(true) argument to while(ContinueSpawning). Once ContinueSpawning is set to false, the the coroutine will not longer enter into the loop and enemy spawning will stop.