2D Laser Properties
Changed 3D laser to 2D laser properties.
Removed following components from existing 3d laser object:
Capsule mesh
Rigidbody (3D)
Collider (3D)

Replaced with:
Sprite Renderer with laser sprite attached
Box 2D collider (is Trigger)
Rigidbody 2D (0 gravity)

The Laser.cs script is attached to the Laser Container object so does not need to be attached to each individual laser.

Created 2 additional laser objects and offset them relative to the center laser so that when Instantiated, the Laser Container will contain 3 individual lasers for each group. There are 3 total groups being instantiated, 1 for the center of the ship, one for the left wing of the ship and one for the right wing of the ship.