Enemy Behavioir Updates
Based on code review with Tom Kessler. He recommended I keep the position resetting of the enemy once it disappears off the viewable screen area, that way, if the player fails to shoot at the enemy, it will continue to spawn, making it more challenging for the player.

I originally took this out because I had already created the spawn manager which handles spawning ememies on its own. But the rationale is that it incentivizes the player to destroy as many enamies as possible.