Tripple Shot Behavior
Created the TrippleShotActive() method inside of Player.cs.
When the TrippleShot object collides with the player, the OnTriggerEnter2D method gets the Player component on the Player Object it collides with. It then calls the player.TrippleShotActive() method on the Player, which sets trippleShotActive bool to true.

Also created a Coroutine and IEnumberator to Wait 5 seconds before setting trippleShotActive bool back to false. That way the tripple shot power up is only active temprarily for 5 seconds.