Speed & Shield Powerups
Created Speed & Shield Powerups.
Speed powerup is set to powerUpID == 1, inside of PowerUp.cs
Shield powerup is set to powerUpID ==2, inside of PowerUp.cs
These ID's are serialized so that they can be set on the inspector for that object.
Changed the TrippleShot sprite to correspond to the Speed and Shield Powerups.
Created StartCoroutine and corresponding IEnumerator methods to add corresponding the powerup behavior.