Mouse Movement Player Controls
Added player input controls using the mouse with Input.GetAxis("Mouse X") and "Mouse Y"
Included a separate if conditional to check if mouse is being used, otherwise, uses the keyboard arrow inputs to move the player around. Also added a sensitivity float variable as a multiplier to the mouse inputs.
Added an exhaust animation for the player ship exhaust.