Shield Sprite Properties
Attached Shield Sprite to Player object.
Created the related animation and resized appropriately.
Set Shield object attached to player to inactive.
Created a reference to the shield object that is accessible inside of the Inspector.
Inside of the ShieldActive() method inside of Player.cs, added the shield.SetActive(true) and inside of the powerdown IEnumerator, sets the shield.SetActive(false) back to false.