Score Text UI
Created UIManager.cs that holds the score text.
Inside of Enemy.cs, calls the player's UpdateScore( ) method and passes the Enemy score value (will depend on the type of enemy).
Created a UpdateScore() method inside Player.cs that takes the Enemy Value integer that's passed by the enemy and increments the player score. Also calls the UpdateScore() method inside of the UIManager.