Player Lives/Health Feature
Created sprites for Player Health indicator.
Placed Canvas Image inside of the Canvas and set the PlayeHealth3 sprite as the default.
Inside of UIManager.cs, created a serialized field that takes in the 4 sprites for Player Health.
Inside of UIManager.cs, created a UpdateHealthSprites(int health) method that updates the livesImage.sprite to correspond to the player health sprite stored in the lives[] array.
Inside of the Player.cs within the TakeDamage( ) method, it calls uiManager.UpdateHealthSprites(health index).