Game Over Feature
Created Game Over feature. Implemented as a UI Image. Created the Game Over sprite using Photoshop and Orbitron font text.
Placed the Image in the scene and set to inactive by default.
Inside of UIManager.cs, created a handle to the Image as a Game Object.
Created a TurnOnGameOverMessage() method that sets the gameOverObject.SetActive(true).
Inside of Player.cs, once Health < 1, calls the uiManager.TurnOnGameOverMessage().