Game Over Image Strobbing Effect
Created the Game Over strobbing effect. This is all implemented inside of the UIManager.cs,
When Game is Over, the "Game Over" message strobes between white and red.
Declared the startColor and endColor variable to Color.white and Color.red, respectively.
Set the strobeSpeed to 1.
Created a gameOver bool.
When the TurnOnGameOverMessage() is called, sets the gameOver bool to true.
Inside of the Update method, if gameOver is true then, gameOverObject.GetComponent<Image>().color = Color.Lerp(colorStart, colorEnd, Mathf.PingPong(Time.time * strobeSpeed, 1))