Game Over Message feature update
Inside UIManager.cs, used IEnumerator to define when the Image turns on and off to add a flashing effect. The Message turns on and off intermittently to create a flashing effect.

gameOverObject.SetActive(true) stays on for .7 seconds before turning off. Stays off for .33 seconds before starting the while loop over again and turning on.

Inside of Update( ) still using the Color.Lerp to give the color strobbing effect.