GameManager Properties
Inside of Build Settings, added the open scene (Game Scene) to the "Scenes in Build" in anticipation of using the Unity Scene Management.

Created GameManager.cs to handle the game state, transitioning from Game Over to starting over after player clicks the Restart button.

In GameManger.cs, created the public method RestartButtonClicked() which calls SceneManager.LoadScene("GameScene").

On the Restart Button OnClick() event, added GameManager in the OnClick() event field and selected the GameManager.RestartButtonClicked() as the method to execute.

Once the Restart button is clicked, the method is called inside of GameManager, and LoadScene method is executed and the scene reloads, restarting the game.