Main Menu Design
Imported Sci-Fi Buttons & Panels Package.
Designed Main Menu UI along with interactable Start button.
Created a MainMenu.cs inside of a MainMenu object in the hierarchy.
Created a transparent button on top of the Rocket logo at the start menu.
The MainMenu calls the OnClick event (StartGame()) when the transparent button is clicked.
When the Start button is clicked, the StartGame() method calls SceneManager.LoadScene("GameScene") to start the game.