Explosion Prefab & Animation
Created enemy explosion prefab and animation.
Inside of Enemy.cs, when collision with Player or laser is detected, Instantiates the explosion prefab with the animation.

The looptime on the animation was disabled.
Also, created an ExplosionController.cs and attached it to the explosion prefab. The ExplosionController simply delays for about .33f (seconds) before self destructing, after it is instantiated.