Asteroid Start Game Features
Created Asteroid game object and its features.
The player must destroy the stationary asteroid at the beginning of the game to start the game.
Created the Asteroid.cs script that controls the asteroids stationary rotation (transform.rotation). Added the Sphere Collider 2D (isTrigger) and Rigidbody 2D (zero gravity) to add the collision elements.

Inside of Asteroid.cs, created the OnTriggerEnter2D(Collision2D other) method to check when  "other" is the laser object. When it's the laser object, destroys the asteroid. Also added a call to a new method (StartSpawning()) inside of the SpawnManager, along with the required SpawnManager handle.

Moved the Start spawning routines that were previously inside of Start() and moved them inside their own method, "StartSpawning()"