Main Level/Game Scene Audio Features
Created Audio Clips for BackgroundWaiting and BackgroundLevel1.
Created handles inside Player.cs for these audio clips.
Defaults to play the BackgroundWaiting and after the asteroid is destroyed, plays the BackgroundLevel1 audio clip.
Set Level1 background clip to looping.

Added an AudioSource to the Player object in the scene.

Fixed a speed boost bug where the speed boost was too sensitive. This was inside of the transform.transpate for the mouse movement detection.