Explosion Properties
Changed the Audio Source Clip in the Explosion and ExplosionBig prefabs to a different explosion audio clip. Changed the clip pitch from 1 to 2.

Inside of Explosion.cs, after the explosion object is instantiated and after a delay of .5f, the Sprite renderer is enabled == false so that it stops rendering the animation, then waits 1f before destroying the explosion object. That way it gives the explosion sound effect attached enough time to play out completely.

ExplosionBig animation was still set to looping so set it to looping false.