Power Up SFX
Added Power Up SFX features to PowerUp.cs.
Added Audio Source to the 3 power up prefabs.
Inside of PowerUp.cs, created handles for the AudioSource and references to the 2 different (temp) audio clips for the power up.
Plays the audio clips when the power up collides with the laser or player.