Mouse sensitivity as a function of Screen.Width
Noticed after switching to "Maximize on Play" for testing that the player movement was much slower than it had been when testing with the non-maximized screen. Double checked the mouse sensitivity setting and it was as expected at 12.

The sensitivity is likely not as responsive due to the varying number of pixels between the two different views.

Tested the mouse sensitivity variable as a function of screen width (Screen.Width). The screen width using the regular display is 407 while the Maximized width was 907. I backed out a screen Width factor as follows: 407 x = 12, which gave x = .0295.

I created a screenWidth factor variable set at .0295f.
At the Start( ) method in Player.cs, I set the mouseSensitivity = screenWidthFactor * Screen.Width.

Tested this new scaling method and it worked pretty well with both resolutions.