Fire Audio Source
Added an Audio Source to the Fire prefab. When the player gets hit by an enemy, when the Fire prefab turns active, will play a Fire SFX.

Created a FireSFX script and attached it to the Fire prefab. The Script gets a reference to the Fire SFX audio Clip and plays it on the Fire Prefab's Audio Source.

At Start (when the prefab is activated by the player's health), set audioSource to enabled to turn it on since it's set to inactive as its default state. Then set the audio source clip to audioClip_Fire and then Play the audio clip.