Updates to Mouse Sensitivity & Pause Game Modal
Mouse sensitivity has discrepancies and player movement sensitivity is not consistent between testing in the IDE and the Windows build.

Wrote 3 if cases to catch the 2 different screen size cases during testing and the target screen size on the Windows build. The screen size during regular testing is 471 pixel width while the screen size on the maximized window during testing is 970.

For testing, if screen size is regular (471 pixel width) mouseSensitivity is not adjusted and simply uses the defualt of mouseSensitivity = 12f. During testing, if decide to use maximized view (width = 970 pixel width), mouseSensitivity of 36 works best for this.

For the Windows build window size (1080x1080) sets mouseSensitivity to the default of 12.

Also created a UI Text that displays test data (mouseSensitivity at Runtime).

Created a Pause Modal and a Resume and Exit button. The Pause Modal activates from the Game Manager after the asteroid is destroyed. This is just a stand in for now until can add more functionality. Currently all it does is it activates the modal. The buttons don't have functionality other than they display a Debug.Log.