In Game Testing Text Configuration
Repositioned the in-game testing text from upper left to lower left of screen so it's less intrusive. Leaving turned on for now, may set inactive later.

Added conditional compile settings for the in-game player sensitivity information display. If compiling for a Windows or WebGL build, then will ignore the if conditionals. If compiling for playing on the editor, will go through the if statements to check screen size and set appropriate mouse sensitivity.