Enemy Variant Properties
Created additional enemy variant, "Enemy_50" and gave it a EnemyValue of 50. Assigns the EnemyValue variable of 50 to this variant and is used when giving points to the player after destroying it. The feature for giving proper number of points (10 points vs. 50 points) was previously built into the Enemy.cs script in previous merges.

This enemy variant is worth 50 points (as opposed to 10) and will be able to fire a laser when inside the game scene, after being spawned.

Inside of Spawn Manager, randomize which variant (variant 0 or variant 1) is spawned, with the variant 0 (Enemy_10) spawn probability of 75% and Enemy_50 with a spawn probability of 25%.

Created Enemy Laser with proton sprite.

Wrote new EnemyLaser script that controls the firing of the photon from the Enemy Variant (Enemy Value 50).

Tested the new enemy variant that it behaves as required. The Enemy Variant with Enemy Value of 50 instantiates with 25% probability and once it reaches a position between y = 7 and 4.3, it fires a photon.

If the photon hits the player, the explosion animation is played and the player takes damage.