Enemy Laser Properties & Mouse Sensitivity Updates
EnemyLaser.cs: previous position where laser destroys self was at y=-2 but that was half way down the visible screen. Reset to y = -10 before enemy laser destroys itself.

Set mouse sensitivity to 4f for when the screen size is between 500 and 1080p, this is the case when running the game inside the editor with maximized view.