Mouse Input for WebGL and Windows Build Properties
Inside of Player.cs, added a compile conditional so that when compiling for Unity Editor play, the game will use mouse input for player control. If using WebGL or Windows build, will use the arrow key for inputs only.

This is for the time being.

Will change this once integrate hand held controller input.