Enemy Wave System Updates
From the Asteroid script, calls a new public method (gameManager.StartWaveTimer()) inside of GameManager that resets the wave timer to zero and sets asteroidDestroyed to true.

The waveTimer was updated so that it is calculated as waveTimer += Time.deltaTime as opposed to Time.time.

Also, the ManageEnemyWaves( ) method doesn't go about setting the wave numbers unless asteroidDestoryed is true.