Laser Detector Feature
Laser Detector feature on Enemy variant Enemy_10.
Attached BoxCollider2D and RigidBody2D to child object on the Enemy_10.
When a laser object is detected, calls a public method on the Enemy that makes the enemy maneuver away from the laser.

NOTE: Make sure that when adding a BoxCollider2D you ALSO add a Rigidbody2D as well. When the laser was colliding with the BoxCollider2D (set as Is Trigger), it was for some reason causing the parent Enemy object to blow up. This stopped once I added the RigidBody2D and the LaserDetector behaved as a trigger.